2010-11-04 22:44 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/hbjson.c   * src/rtl/hblpp.c     ! Silenced msvc64 warnings.
  * utils/hbrun/hbrun.prg
    * Moved creation of header hash to separate function.
    + Added HBRUN_NOHEAD envver to disable embedded headers.
      Any non-empty value will have this effect.

  * utils/hbrun/Makefile
    + Enabled embedded headers.

  * contrib/make.hbs
  * config/postinst.hbs
  * bin/hbxpatch.hbs
    ! Resolved TOFIX notes: now the headers can be used
      just like in any .prg, and no need to locally #define
      core constants.
      Also hbqt_hbmk2_plugin.hbs should work now in stdalone 
      mode, with all the OOP code.

+++ harbour/src/rtl/hbjson.c
@@ -498,7 +498,7 @@ static const char * _hb_jsonDecode( const char * szSource, PHB_ITEM pValue )
       {
          double mult = 1;
 
-         dblValue = nValue;
+         dblValue = ( double ) nValue;
          fDbl = HB_TRUE;
          szSource++;
          while( *szSource >= '0' && *szSource <= '9' )
@@ -526,7 +526,7 @@ static const char * _hb_jsonDecode( const char * szSource, PHB_ITEM pValue )
          }
          if( ! fDbl )
          {
-            dblValue = nValue;
+            dblValue = ( double ) nValue;
             fDbl = HB_TRUE;
          }
          dblValue *= pow( 10.0, ( double ) ( fNegExp ? -iExp : iExp ) );
